# server/openapi.yaml
openapi: 3.0.1
info:
  title: Accountable Server
  version: 0.8.0
  description: This is the Accountable Server. All the server does is store data given to it by the client. The server controls access to the data with basic password-based auth with optional time-based multi-factor auth. The Accountable client does most of the hard work of encrypting the user's data. You can read Accountable's source code on [GitHub](https://github.com/AverageHelper/accountable-svelte).
  license:
    name: GNU General Public License v3.0
    url: https://github.com/AverageHelper/accountable-svelte/blob/main/LICENSE
...
servers:
  - url: https://api.accountable.average.name
...
paths:
...
  "/v0/totp/secret":
    get:
      tags:
        - auth
      summary: Generates a new TOTP secret for the user. If the user has already enabled 2FA, then a 409 status is returned.
      operationId: get-totp-secret
      security:
        - jwt: []
      responses:
        200:
          description: Got a new secret.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Success!"
                  secret:
                    type: string
                    description: A TOTP configuration URI, to be displayed to the user as, for example, a QR code. The user should keep this value safe, as it is the only way to generate time-based authentication codes.
        403:
          description: The caller did not send valid session cookies.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - "expired-token"
                      - "missing-token"
        409:
          description: The user already has a TOTP secret set. We should not override it.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - "totp-conflict"
    delete:
      tags:
        - auth
      summary: Disables the user's TOTP requirement, and deletes the stored TOTP secret.
      operationId: delete-totp-secret
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - password
                - token
              type: object
              properties:
                password:
                  type: string
                  description: The passphrase to access the account.
                token:
                  type: string
                  description: The current TOTP to access the account.
      responses:
        200:
          description: The secret is gone.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Success!"
        403:
          description: The caller did not send valid session cookies, or the password was invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - "expired-token"
                      - "missing-token"
                      - "wrong-credentials"
                      - "wrong-mfa-credentials"
...
components:
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT